frontend/handwriting_recognition.py
import streamlit as st
import os
import tempfile
from PIL import Image
import numpy as np
from api_layer.llm_interface import GeminiInterface

def render_handwriting():
    """Render the handwriting recognition UI component."""
    st.header("Handwriting Recognition")
    st.write("Upload images of handwritten notes for digital conversion.")
    
    # Initialize session state variables if they don't exist
    if "recognized_text" not in st.session_state:
        st.session_state.recognized_text = ""
    if "llm" not in st.session_state:
        try:
            st.session_state.llm = GeminiInterface()
        except Exception as e:
            st.error(f"Error initializing Gemini AI: {str(e)}")
            st.info("Please check your API key in the .env file.")
            return
    
    # File uploader for handwritten images
    uploaded_file = st.file_uploader("Upload an image of handwritten text", type=["jpg", "jpeg", "png"])
    
    if uploaded_file is not None:
        # Display the uploaded image
        image = Image.open(uploaded_file)
        st.image(image, caption="Uploaded Image", use_column_width=True)
        
        # Process button
        if st.button("Recognize Text", key="recognize_button"):
            with st.spinner("Processing image..."):
                try:
                    # Save the uploaded image to a temporary file
                    temp_dir = tempfile.mkdtemp()
                    temp_file_path = os.path.join(temp_dir, uploaded_file.name)
                    with open(temp_file_path, "wb") as f:
                        f.write(uploaded_file.getbuffer())
                    
                    # Placeholder for actual OCR implementation
                    # In a real application, you would use an OCR service or library
                    # For now, we'll use a placeholder message
                    
                    # Simulate OCR with a placeholder message
                    recognized_text = "[This is a placeholder for handwriting recognition. In a real application, this would use an OCR service or library.]\n\nPlease upload your handwritten notes to get actual text recognition."
                    
                    # Store in session state
                    st.session_state.recognized_text = recognized_text
                    
                    # Display the recognized text
                    st.subheader("Recognized Text")
                    st.text_area("Extracted Text", recognized_text, height=300)
                    
                    # Post-processing options
                    st.subheader("Post-Processing Options")
                    post_process_tabs = st.tabs(["Edit", "Summarize", "Format"])
                    
                    with post_process_tabs[0]:
                        edited_text = st.text_area("Edit the recognized text:", recognized_text, height=300, key="edit_text")
                        
                        if st.button("Save Edits", key="save_edits_button"):
                            st.session_state.recognized_text = edited_text
                            st.success("Edits saved!")
                    
                    with post_process_tabs[1]:
                        if st.button("Generate Summary", key="summarize_button"):
                            with st.spinner("Generating summary..."):
                                try:
                                    # Use Gemini to summarize the text
                                    prompt = f"Summarize the following text extracted from handwritten notes:\n\n{st.session_state.recognized_text}"
                                    
                                    summary = st.session_state.llm.generate_text(prompt, temperature=0.3)
                                    
                                    st.subheader("Summary")
                                    st.write(summary)
                                except Exception as e:
                                    st.error(f"Error generating summary: {str(e)}")
                    
                    with post_process_tabs[2]:
                        format_option = st.selectbox(
                            "Format As",
                            ["Plain Text", "Markdown", "Bullet Points", "Numbered List"]
                        )
                        
                        if st.button("Format Text", key="format_button"):
                            with st.spinner("Formatting text..."):
                                try:
                                    # Use Gemini to format the text
                                    prompt = f"Format the following text as {format_option}:\n\n{st.session_state.recognized_text}"
                                    
                                    formatted_text = st.session_state.llm.generate_text(prompt, temperature=0.3)
                                    
                                    st.subheader("Formatted Text")
                                    st.markdown(formatted_text)
                                except Exception as e:
                                    st.error(f"Error formatting text: {str(e)}")
                    
                    # Export options
                    st.subheader("Export Options")
                    export_format = st.selectbox(
                        "Export Format",
                        ["TXT", "MD", "PDF", "DOCX"]
                    )
                    
                    if st.button("Export", key="export_button"):
                        # Placeholder for export functionality
                        st.info(f"Export to {export_format} would be implemented in a real application.")
                        st.download_button(
                            label=f"Download as {export_format}",
                            data=st.session_state.recognized_text,
                            file_name=f"recognized_text.{export_format.lower()}",
                            mime="text/plain"
                        )
                
                except Exception as e:
                    st.error(f"Error processing image: {str(e)}")
    else:
        st.info("Please upload an image of handwritten text to get started.")
        
        # Example images
        st.subheader("Example")
        st.write("Here's an example of how this feature works:")
        
        col1, col2 = st.columns(2)
        with col1:
            st.markdown("**1. Upload an image of handwritten notes**")
            st.markdown("**2. Click 'Recognize Text' to process the image**")
            st.markdown("**3. Edit, summarize, or format the extracted text**")
            st.markdown("**4. Export the text in your preferred format**")
        
        with col2:
            # Display a placeholder example image
            example_image = np.zeros((300, 400, 3), dtype=np.uint8)
            example_image.fill(240)  # Light gray background
            st.image(example_image, caption="Example Handwritten Notes", use_column_width=True)