requirements.txt
# StudyMateZ dependencies

# Core dependencies
streamlit>=1.22.0
python-dotenv>=1.0.0

# AI and NLP
google-generativeai>=0.3.0
langchain>=0.0.267
langchain-google-genai>=0.0.5
nltk>=3.8.1

# PDF processing
pymupdf>=1.22.5
pdfplumber>=0.9.0

# Vector database
chromadb>=0.4.13

# Data processing and visualization
pandas>=2.0.3
matplotlib>=3.7.2
altair>=5.0.1
plotly>=5.15.0
networkx>=3.1

# Audio processing
pydub>=0.25.1
numpy>=1.24.3
pillow>=9.5.0

# Utilities
tqdm>=4.65.0